# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ObjectId": {
				"type": "string",
				"description": "A class representation of the BSON ObjectId type."
			},
			"Maybe_number-Array-Array-Array_": {
				"items": {
					"items": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"type": "array"
				},
				"type": "array",
				"nullable": true
			},
			"Maybe_string_": {
				"type": "string",
				"nullable": true
			},
			"EntityType": {
				"type": "string",
				"enum": [
					"animal",
					"plant"
				]
			},
			"Maybe_number_": {
				"type": "number",
				"format": "double",
				"nullable": true
			},
			"LocationSource": {
				"type": "string",
				"enum": [
					"custom",
					"search"
				]
			},
			"Maybe_LocationSource_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/LocationSource"
					}
				],
				"nullable": true
			},
			"SearchRecordStatus": {
				"type": "string",
				"enum": [
					"COMPLETE",
					"READY",
					"SCRAPING"
				]
			},
			"Pick_SearchRecordDocument.Exclude_keyofSearchRecordDocument._id__": {
				"properties": {
					"userIds": {
						"items": {
							"$ref": "#/components/schemas/ObjectId"
						},
						"type": "array"
					},
					"taxonKeys": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"SearchRecord"
						]
					},
					"boundingPolyCoords": {
						"$ref": "#/components/schemas/Maybe_number-Array-Array-Array_"
					},
					"commonName": {
						"$ref": "#/components/schemas/Maybe_string_"
					},
					"createdTimestamp": {
						"type": "number",
						"format": "double"
					},
					"entityType": {
						"$ref": "#/components/schemas/EntityType"
					},
					"lastRanTimestamp": {
						"$ref": "#/components/schemas/Maybe_number_"
					},
					"locationName": {
						"$ref": "#/components/schemas/Maybe_string_"
					},
					"locationSource": {
						"$ref": "#/components/schemas/Maybe_LocationSource_"
					},
					"occurrencesOffset": {
						"type": "number",
						"format": "double"
					},
					"scientificName": {
						"$ref": "#/components/schemas/Maybe_string_"
					},
					"status": {
						"$ref": "#/components/schemas/SearchRecordStatus"
					},
					"totalOccurrences": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"createdTimestamp",
					"entityType",
					"occurrencesOffset",
					"status",
					"totalOccurrences"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_SearchRecordDocument._id_": {
				"$ref": "#/components/schemas/Pick_SearchRecordDocument.Exclude_keyofSearchRecordDocument._id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"SearchRecordSummary": {
				"allOf": [
					{
						"properties": {
							"id": {
								"type": "string"
							}
						},
						"required": [
							"id"
						],
						"type": "object"
					},
					{
						"$ref": "#/components/schemas/Omit_SearchRecordDocument._id_"
					}
				]
			},
			"PlantSearchLocationParams": {
				"properties": {
					"boundingPolyCoords": {
						"items": {
							"items": {
								"items": {
									"type": "number",
									"format": "double"
								},
								"type": "array"
							},
							"type": "array"
						},
						"type": "array"
					},
					"locationSource": {
						"type": "string",
						"enum": [
							"search",
							"custom"
						]
					},
					"locationName": {
						"type": "string"
					}
				},
				"required": [
					"boundingPolyCoords",
					"locationSource",
					"locationName"
				],
				"type": "object"
			},
			"EntityNameParams": {
				"anyOf": [
					{
						"properties": {
							"commonName": {
								"type": "string"
							}
						},
						"required": [
							"commonName"
						],
						"type": "object"
					},
					{
						"properties": {
							"scientificName": {
								"type": "string"
							}
						},
						"required": [
							"scientificName"
						],
						"type": "object"
					}
				]
			},
			"EntitySearchParams": {
				"properties": {
					"entityType": {
						"$ref": "#/components/schemas/EntityType"
					},
					"entityName": {
						"$ref": "#/components/schemas/EntityNameParams"
					},
					"location": {
						"$ref": "#/components/schemas/PlantSearchLocationParams"
					}
				},
				"required": [
					"entityType"
				],
				"type": "object"
			},
			"Maybe_string-Array_": {
				"items": {
					"type": "string"
				},
				"type": "array",
				"nullable": true
			},
			"Maybe_number-Array_": {
				"items": {
					"type": "number",
					"format": "double"
				},
				"type": "array",
				"nullable": true
			},
			"PlantArrayValuesDocument": {
				"properties": {
					"bloomColors": {
						"$ref": "#/components/schemas/Maybe_string-Array_"
					},
					"bloomTimes": {
						"$ref": "#/components/schemas/Maybe_string-Array_"
					},
					"habitats": {
						"$ref": "#/components/schemas/Maybe_string-Array_"
					},
					"hardiness": {
						"$ref": "#/components/schemas/Maybe_number-Array_"
					},
					"lightLevels": {
						"$ref": "#/components/schemas/Maybe_string-Array_"
					},
					"soilTypes": {
						"$ref": "#/components/schemas/Maybe_string-Array_"
					}
				},
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "hotplants-server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/searchRecord": {
			"post": {
				"operationId": "GetSearchRecord",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SearchRecordSummary"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EntitySearchParams"
							}
						}
					}
				}
			}
		},
		"/runSearch/{searchRecordId}": {
			"get": {
				"operationId": "RunSearch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SearchRecordSummary"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "Initiate a new search of occurrences from GBIF and combine with PFAF data.",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "searchRecordId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/plants/filterValues": {
			"get": {
				"operationId": "GetFilterValues",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlantArrayValuesDocument"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}